Large class refactor, rename task to hook
