+++ Sources/InterposeKit/ObjectHook.swift
@@ -16,50 +16,37 @@ internal enum ObjCMethodEncoding {
     }
 }
 
-/// A task represents a hook to an instance method of a single object and stores both the original and new implementation.
-final public class ObjectTask: ValidatableTask {
+/// A hook to an instance method of a single object, stores both the original and new implementation.
+/// TODO: Multiple hooks for one object
+final class ObjectHook: InternalHookable {
     public let `class`: AnyClass
-    public let object: AnyObject
     public let selector: Selector
-    public private(set) var origIMP: IMP? // fetched at apply time, changes late, thus class requirement
+    public internal(set) var origIMP: IMP? // fetched at apply time, changes late, thus class requirement
     public private(set) var replacementIMP: IMP! // else we validate init order
-    public private(set) var state = Interpose.State.prepared
+    public internal(set) var state = Interpose.State.prepared
 
-    // Subclass that we create on the fly
+    public let object: AnyObject
+    /// Subclass that we create on the fly
     var dynamicSubclass: AnyClass?
 
-    /// Initialize a new task to interpose an instance method.
-    public init(object: AnyObject, selector: Selector, implementation: (Task) -> Any) throws {
-        self.selector = selector
+    /// Initialize a new hook to interpose an instance method.
+    public init(object: AnyObject, selector: Selector, implementation: (Hookable) -> Any) throws {
         self.object = object
+        self.selector = selector
         self.class = type(of: object)
         // Check if method exists
         try validate()
         replacementIMP = imp_implementationWithBlock(implementation(self))
     }
 
-    class KVOObserver: NSObject {
-        @objc var objectToObserve: AnyObject
-        var observation: NSKeyValueObservation?
+//    /// Release the hook block if possible.
+//    public override func cleanup() {
+//        // TODO: remove subclass!
+//        super.cleanup()
+//    }
 
-        init(object: AnyObject) {
-            objectToObserve = object
-            super.init()
-
-            // Can't use modern syntax cause https://bugs.swift.org/browse/SR-12944
-            objectToObserve.addObserver(self, forKeyPath: "description", options: .new, context: nil)
-        }
-    }
-
-    // Before creating our subclass, we trigger KVO.
-    // KVO also creates a subclass at runtime. If we do this prior, then KVO fails.
-    // If KVO runs prior, and then we sub-subclass, everything works.
-    var kvoObserver: KVOObserver?
-    private func registerKVO() {
-        kvoObserver = KVOObserver(object: object)
-    }
-
-    private func createSubclass() throws -> AnyClass {
+    /// Creates a unique dynamic subclass of the current object
+    private func createDynamicSubclass() throws -> AnyClass {
         let perceivedClass: AnyClass = `class`
         let className = NSStringFromClass(perceivedClass)
         // Right now we are wasteful. Might be able to optimize for shared IMP?
@@ -130,47 +117,8 @@ final public class ObjectTask: ValidatableTask {
         }
         class_addMethod(subclass, self.selector, imp_implementationWithBlock(block), typeEncoding)
     }
-    
-    /// Validate that the selector exists on the active class.
-    @discardableResult public func validate(expectedState: Interpose.State = .prepared) throws -> Method {
-        // We need to validate on class, not the subclass
-        guard let method = class_getInstanceMethod(`class`, selector) else { throw Interpose.Error.methodNotFound }
-        guard state == expectedState else { throw Interpose.Error.invalidState }
-        return method
-    }
 
-    public func apply() throws {
-        try execute(newState: .interposed) { try replaceImplementation() }
-    }
-
-    public func revert() throws {
-        try execute(newState: .prepared) { try resetImplementation() }
-    }
-
-    /// Release the hook block if possible.
-    public func cleanup() {
-        switch state {
-        case .prepared:
-            Interpose.log("Releasing -[\(`class`).\(selector)] IMP: \(replacementIMP!)")
-            imp_removeBlock(replacementIMP)
-        case .interposed:
-            Interpose.log("Keeping -[\(`class`).\(selector)] IMP: \(replacementIMP!)")
-        case let .error(error):
-            Interpose.log("Leaking -[\(`class`).\(selector)] IMP: \(replacementIMP!) due to error: \(error)")
-        }
-    }
-
-    private func execute(newState: Interpose.State, task: () throws -> Void) throws {
-        do {
-            try task()
-            state = newState
-        } catch let error as Interpose.Error {
-            state = .error(error)
-            throw error
-        }
-    }
-
-    private func replaceImplementation() throws {
+    func replaceImplementation() throws {
         let method = try validate()
 
         // Register a KVO to work around any KVO issues with opposite order
@@ -178,7 +126,7 @@ final public class ObjectTask: ValidatableTask {
 
         // Register subclass at runtime if we haven't already
         if dynamicSubclass == nil {
-            dynamicSubclass = try createSubclass()
+            dynamicSubclass = try createDynamicSubclass()
         }
         // Add empty trampoline that we then replace the IMP!
         addSuperTrampolineMethod(subclass: dynamicSubclass!, method: method)
@@ -188,25 +136,60 @@ final public class ObjectTask: ValidatableTask {
         Interpose.log("Swizzled -[\(`class`).\(selector)] IMP: \(origIMP!) -> \(replacementIMP!)")
     }
 
-    private func resetImplementation() throws {
+    func resetImplementation() throws {
         let method = try validate(expectedState: .interposed)
         precondition(origIMP != nil)
-        precondition(dynamicSubclass != nil)
+        guard let dynamicSubclass = self.dynamicSubclass else { preconditionFailure("No dynamic subclass set") }
 
-        let previousIMP = class_replaceMethod(dynamicSubclass!, selector, origIMP!, method_getTypeEncoding(method))
+        let previousIMP = class_replaceMethod(dynamicSubclass, selector, origIMP!, method_getTypeEncoding(method))
         guard previousIMP == replacementIMP else { throw Interpose.Error.unexpectedImplementation }
         Interpose.log("Restored -[\(`class`).\(selector)] IMP: \(origIMP!)")
+
+        // Restore the original class of the object
+        // TODO: Does this include the KVO'ed subclass?
+        object_setClass(object, `class`)
+
+        // Dispose of the custom dynamic subclass
+        objc_disposeClassPair(dynamicSubclass)
+        self.dynamicSubclass = nil
+
+        // Remove KVO after restoring class as last step.
+        deregisterKVO()
     }
 
-    public func callAsFunction<U>(_ type: U.Type) -> U {
-        unsafeBitCast(origIMP, to: type)
+// MARK: KVO Helper
+
+    var kvoObserver: KVOObserver?
+
+    class KVOObserver: NSObject {
+        @objc var objectToObserve: AnyObject
+        var observation: NSKeyValueObservation?
+
+        init(object: AnyObject) {
+            objectToObserve = object
+            super.init()
+
+            // Can't use modern syntax cause https://bugs.swift.org/browse/SR-12944
+            objectToObserve.addObserver(self, forKeyPath: "description", options: .new, context: nil)
+        }
+    }
+
+    // Before creating our subclass, we trigger KVO.
+    // KVO also creates a subclass at runtime. If we do this prior, then KVO fails.
+    // If KVO runs prior, and then we sub-subclass, everything works.
+    private func registerKVO() {
+        kvoObserver = KVOObserver(object: object)
+    }
+
+    private func deregisterKVO() {
+        kvoObserver = nil
     }
 }
 
 #if DEBUG
-extension ObjectTask: CustomDebugStringConvertible {
+extension ObjectHook: CustomDebugStringConvertible {
     public var debugDescription: String {
-        return "\(selector) -> \(String(describing: origIMP))"
+        return "\(selector) of \(object) -> \(String(describing: origIMP))"
     }
 }
 #endif
